updating read me and cleaning code

+++ bamazon.sql
@@ -9,7 +9,7 @@ id INT AUTO_INCREMENT,
 product_name VARCHAR(1000) NULL,
 department_name VARCHAR(1000) NULL,
 price DECIMAL(10, 2),
-stock_quantity DECIMAL(10, 2),
+stock_quantity INTEGER(10),
 PRIMARY KEY(id)
 );
